+++ src/decorators.ts
@@ -20,7 +20,11 @@ export function customElement(tagname?: string) {
   // TODO Investigate narrowing down the type of clazz.
   return (clazz: any) => {
     // TODO(justinfagnani): we could also use the kebab-cased class name
-    tagname = clazz.is = clazz.is || tagname;
+    if (clazz.is) {
+      tagname = clazz.is;
+    } else {
+      clazz.is = tagname;
+    }
     window.customElements.define(tagname!, clazz);
   }
 }
@@ -38,6 +42,38 @@ export interface PropertyOptions {
   notify?: boolean;
   reflectToAttribute?: boolean;
   readOnly?: boolean;
+  computed?: string;
+  observer?: string|((val: any, old: any) => void);
+}
+
+function createProperty(
+    proto: any, name: string, options?: PropertyOptions): void {
+  const notify = options && options.notify || false;
+  const reflectToAttribute = options && options.reflectToAttribute || false;
+  const readOnly = options && options.readOnly || false;
+  const computed = options && options.computed || '';
+  const observer = options && options.observer || '';
+
+  let type;
+  if (options && options.hasOwnProperty('type')) {
+    type = options.type;
+  } else if (
+      (window as any).Reflect && Reflect.hasMetadata && Reflect.getMetadata &&
+      Reflect.hasMetadata('design:type', proto, name)) {
+    type = Reflect.getMetadata('design:type', proto, name);
+  } else {
+    console.error(
+        'A type could not be found for ${propName}. ' +
+        'Set a type or configure Metadata Reflection API support.');
+  }
+
+  if (!proto.constructor.hasOwnProperty('properties')) {
+    Object.defineProperty(proto.constructor, 'properties', {value: {}});
+  }
+
+  const finalOpts: PropertyOptions =
+      {type, notify, reflectToAttribute, readOnly, computed, observer};
+  proto.constructor.properties[name] = finalOpts;
 }
 
 /**
@@ -48,33 +84,7 @@ export interface PropertyOptions {
  */
 export function property(options?: PropertyOptions) {
   return (proto: any, propName: string): any => {
-    const notify: boolean = options && options.notify || false;
-    const reflectToAttribute: boolean =
-        options && options.reflectToAttribute || false;
-    const readOnly: boolean = options && options.readOnly || false;
-
-    let type;
-    if (options && options.hasOwnProperty('type')) {
-      type = options.type;
-    } else if (
-        (window as any).Reflect && Reflect.hasMetadata && Reflect.getMetadata &&
-        Reflect.hasMetadata('design:type', proto, propName)) {
-      type = Reflect.getMetadata('design:type', proto, propName);
-    } else {
-      console.error(
-          'A type could not be found for ${propName}. ' +
-          'Set a type or configure Metadata Reflection API support.');
-    }
-
-    if (!proto.constructor.hasOwnProperty('properties')) {
-      proto.constructor.properties = {};
-    }
-    proto.constructor.properties[propName] = {
-      type,
-      notify,
-      reflectToAttribute,
-      readOnly,
-    };
+    createProperty(proto, propName, options);
   }
 }
 
@@ -94,6 +104,30 @@ export function observe(targets: string|string[]) {
     }
     proto.constructor.observers.push(`${propName}(${targetString})`);
   }
+}
+
+/**
+ * A TypeScript accessor decorator factory that causes the decorated getter to
+ * be called when a set of dependencies change. The arguments of this decorator
+ * should be paths of the data dependencies as described
+ * [here](https://www.polymer-project.org/2.0/docs/devguide/observers#define-a-computed-property)
+ * The decorated getter should not have an associated setter.
+ *
+ * This function must be invoked to return a decorator.
+ */
+export function computed<T = any>(...targets: (keyof T)[]) {
+  return (proto: any,
+          propName: string,
+          descriptor: PropertyDescriptor): void => {
+    const fnName = `__compute${propName}`;
+
+    Object.defineProperty(proto, fnName, {value: descriptor.get});
+
+    descriptor.get = undefined;
+
+    createProperty(
+        proto, propName, {computed: `${fnName}(${targets.join(',')})`});
+  };
 }
 
 /**
@@ -141,3 +175,22 @@ function _query(
     });
   };
 }
+
+/**
+ * A TypeScript property decorator factory that causes the decorated method to
+ * be called when a imperative event is fired on the targeted element. `target`
+ * can be either a single element by id or element.
+ *
+ * You must apply the supplied DeclarativeEventListeners mixin to your element
+ * class for this decorator to function.
+ *
+ * https://www.polymer-project.org/2.0/docs/devguide/events#imperative-listeners
+ *
+ * @param eventName A string representing the event type to listen for
+ * @param target A single element by id or EventTarget to target
+ */
+export const listen = (eventName: string, target: string|EventTarget) =>
+    (proto: any, methodName: string) => {
+      proto.constructor._addDeclarativeEventListener(
+          target, eventName, proto[methodName]);
+    };