+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.1.2] - 2018-01-01
+- Added `observer` and `computed` properties to the `@property` decorator options.
+- Added `@listen` decorator and `DeclarativeEventListeners` mixin.
+- Fixed bug where use of `Polymer.mixinBehaviors` with `@property` or `@computed` would throw an exception relating to the definition of the element's "properties" property.
+
 ## [0.1.1] - 2017-09-28
 - Metadata Reflection polyfill is no longer a bower dependency.